Refactor LastSavedText in MeasureWrapper
