+++ services/ui-src/src/components/MeasureButtons/index.tsx
@@ -1,23 +1,25 @@
 import * as CUI from "@chakra-ui/react";
 import * as QMR from "components";
-import { FaCheckCircle } from "react-icons/fa";
 import { useUser } from "hooks/authHooks";
 interface Props {
   handleSave: () => void;
-  lastSavedText?: string;
+  lastAltered?: number;
+  isLoading?: boolean;
 }
 
-export const MeasureButtons = ({ handleSave, lastSavedText }: Props) => {
+export const MeasureButtons = ({
+  handleSave,
+  lastAltered,
+  isLoading,
+}: Props) => {
   const { isStateUser } = useUser();
-
-  const showCheck = lastSavedText?.toLowerCase() === "saved moments ago";
 
   return (
     <CUI.Stack width="15rem">
       <CUI.HStack pl="1">
         <QMR.ContainedButton
-          disabledStatus={!isStateUser}
-          buttonText={"Save"}
+          disabledStatus={!isStateUser || isLoading}
+          buttonText={isLoading ? "Saving" : "Save"}
           buttonProps={{
             minWidth: "10rem",
             colorScheme: "blue",
@@ -26,18 +28,11 @@ export const MeasureButtons = ({ handleSave, lastSavedText }: Props) => {
           onClick={handleSave}
         />
       </CUI.HStack>
-      <CUI.Flex mb={{ base: "1", lg: "0" }} data-testid="last-saved-text">
-        {showCheck && (
-          <CUI.Box mt="1">
-            <FaCheckCircle data-testid="circle-check-icon" />
-          </CUI.Box>
-        )}
-        {lastSavedText && (
-          <CUI.Text ml={showCheck ? 2 : 0} fontSize="sm">
-            {lastSavedText}
-          </CUI.Text>
-        )}
-      </CUI.Flex>
+      {lastAltered && (
+        <CUI.Flex mb={{ base: "1", lg: "0" }} data-testid="last-saved-text">
+          <QMR.LastSavedText lastAltered={lastAltered} />
+        </CUI.Flex>
+      )}
     </CUI.Stack>
   );
 };
