+++ services/ui-src/src/views/Qualifiers/ChildMedicaid/index.tsx
@@ -7,25 +7,6 @@ import { useParams, useNavigate } from "react-router-dom";
 import { useUpdateMeasure, useGetMeasure } from "hooks/api";
 import { CoreSetAbbr, MeasureStatus } from "types";
 import { useQueryClient } from "react-query";
-import { formatDistanceToNow } from "date-fns";
-import { FaCheckCircle } from "react-icons/fa";
-
-const LastSavedText = ({ lastAltered }: { lastAltered?: number }) => {
-  if (!lastAltered) return null;
-  const lastAlteredText = formatDistanceToNow(new Date(lastAltered), {
-    addSuffix: true,
-  });
-  return (
-    <CUI.Flex justifyContent="center">
-      <CUI.Box mt="1">
-        <FaCheckCircle />
-      </CUI.Box>
-      <CUI.Text ml="2" fontSize="sm">
-        {`Submitted ${lastAlteredText}`}
-      </CUI.Text>
-    </CUI.Flex>
-  );
-};
 
 export const CCSMQualifiers = () => {
   const { state, year } = useParams();
@@ -106,7 +87,7 @@ export const CCSMQualifiers = () => {
       ]}
       buttons={
         data?.Item?.data && (
-          <LastSavedText lastAltered={data?.Item.lastAltered} />
+          <QMR.LastSavedText lastAltered={data?.Item.lastAltered} />
         )
       }
     >
